# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ForkBody": {
        "additionalProperties": false,
        "description": "What to call the forked goal.",
        "properties": {
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "title": "ForkBody",
        "type": "object"
      },
      "GoalBundleBody": {
        "additionalProperties": false,
        "description": "An imported bundle, plus the slug to import it under.",
        "properties": {
          "bundle": {
            "additionalProperties": true,
            "title": "Bundle",
            "type": "object"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "required": [
          "bundle"
        ],
        "title": "GoalBundleBody",
        "type": "object"
      },
      "GoalPackBody": {
        "additionalProperties": false,
        "description": "A goal pack as a request body: its ``goal.json`` and its task prompt records.\n\nTwo fields rather than one nested document, because they are two different kinds of\nthing: the goal is a rubric, and each task is an ADR-0012 prompt record validated by the\nprompt loader rather than by this model.",
        "properties": {
          "goal": {
            "additionalProperties": true,
            "title": "Goal",
            "type": "object"
          },
          "tasks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "goal"
        ],
        "title": "GoalPackBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FreeWeight",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/goals": {
      "get": {
        "description": "List the goal packs this installation can run.\n\nA pack that cannot be parsed at all is omitted rather than failing the listing: nine working\ngoals must not be hidden by a tenth with a typo, and ``POST /goals/{slug}/validate`` is where\nthe tenth is explained.",
        "operationId": "list_goals_endpoint_api_v1_goals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Goals Endpoint Api V1 Goals Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Every installed goal",
        "tags": [
          "goals"
        ]
      },
      "post": {
        "description": "Write a new goal pack and return it with its lint findings.\n\nFindings never block creation; an ``error`` finding refuses through ``GOAL_INVALID`` because\nsuch a pack could not be run at all.",
        "operationId": "create_goal_endpoint_api_v1_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalPackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a goal",
        "tags": [
          "goals"
        ]
      }
    },
...
    "/api/v1/goals/import": {
      "post": {
        "description": "Import a portable goal bundle.\n\nEverything is validated against the bundle in memory — size, member names, hash, slug\navailability — before a single file is written, and an import never overwrites an existing\ngoal in place (spec §14).",
        "operationId": "import_goal_endpoint_api_v1_goals_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalBundleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import a bundle",
        "tags": [
          "goals"
        ]
      }
    },
...
    "/api/v1/goals/starters/{key}/fork": {
      "post": {
        "description": "Fork one starter into the user's goals root.\n\nThe copy is an ordinary directory of JSON the user owns: editable in an editor, diffable in\ngit, portable to another machine. It is badged ``unforked`` until its criteria or its tasks\nare edited — a voice measured on somebody else's prompts is not the user's voice, and the\nbadge is what stops a starter quietly becoming a default.\n\nRaises:\n    StarterNotFound: No starter has that key.\n    GoalSlugCollision: A goal with that slug already exists.",
        "operationId": "fork_starter_endpoint_api_v1_goals_starters__key__fork_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ForkBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy a starter into your own goals",
        "tags": [
          "goals"
        ]
      }
    },
    "/api/v1/goals/{slug}": {
      "delete": {
        "description": "Preview or perform a goal's deletion.\n\nA bare ``DELETE`` previews. The preview names the two things that are expensive to lose: the\nruns it orphans, and the grades the user produced by hand.",
        "operationId": "delete_goal_endpoint_api_v1_goals__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Goal Endpoint Api V1 Goals  Slug  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a goal, preview first",
        "tags": [
          "goals"
        ]
      },
      "get": {
        "description": "Return one goal as loaded: its lint findings, its pack and its calibration report.\n\n``pack`` is ``goal.json`` and the task records exactly as they are on disk — the body ``PUT``\ntakes — so an editor starts from the documents rather than from this summary, which would drop\nwhatever the summary does not carry. ``calibration`` is the stored report, or ``null``.",
        "operationId": "get_goal_endpoint_api_v1_goals__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Goal Endpoint Api V1 Goals  Slug  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One goal",
        "tags": [
          "goals"
        ]
      },
      "put": {
        "description": "Replace one goal, reporting what the change separates.\n\nThe response carries ``hash_change``: the previous and current ``goal_hash``, whether they\ndiffer, which parts of the measurement-defining document moved, and how many existing runs\nthe previous hash holds — the runs a separating change leaves behind.\n\n``?dry_run=true`` builds and validates the replacement and then discards it, so the same\nreport can be shown **before** the change is applied. That is the order acceptance criterion 4\nasks for: a user is told what an edit would separate while they can still decide not to make\nit.",
        "operationId": "replace_goal_endpoint_api_v1_goals__slug__put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalPackBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replace Goal Endpoint Api V1 Goals  Slug  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace a goal",
        "tags": [
          "goals"
        ]
      }
    },
    "/api/v1/goals/{slug}/bundle": {
      "get": {
        "description": "Return one goal as the bundle ``freeweight goals export`` writes — the round-trip form.\n\nEvery file of the pack, hash-pinned, which is what ``POST /goals/import`` reads back. Byte for\nbyte the CLI's document (:func:`~freeweight.services.goals.bundle_text`), so the two cannot\ncome to differ, and served as an attachment because it is a file to keep.",
        "operationId": "bundle_endpoint_api_v1_goals__slug__bundle_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the portable bundle",
        "tags": [
          "goals"
        ]
      }
    },
...
    "/api/v1/goals/{slug}/calibration/report/export": {
      "get": {
        "description": "Return one goal's calibration as a ``benchmark.calibration_report`` SetSpec envelope.\n\nThe contract form of the report ``GET /goals/{slug}/calibration/report`` returns: the same\nmeasurement, in the schema spec §7.3 names, so a consumer can read the agreement figures\nwithout knowing FreeWeight's own field names. Every coefficient carries its ``n_holdout``,\nwhich SetSpec's own model requires.\n\nRaises:\n    ExportRefused: The goal has never been calibrated. An uncalibrated goal and an empty\n        report are different things, and this refuses rather than conflating them.",
        "operationId": "export_calibration_report_endpoint_api_v1_goals__slug__calibration_report_export_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export as benchmark.calibration_report",
        "tags": [
          "goals"
        ]
      }
    },
...
  }
}